refactor(payment): create a hotel options function
