+++ src/pages/Dashboard/Payment/index.js
@@ -8,13 +8,12 @@ import { FaCheckCircle } from 'react-icons/fa';
 import CredCard from '../../../hooks/useCredCard';
 import usePaidTicket from '../../../hooks/api/usePaidTicket';
 
-import Button from '../../../components/Form/Button.js';
 import useTicketTypes from '../../../hooks/api/useTicketTypes.js';
-import useCreateTicket from '../../../hooks/api/useCreateTicket.js';
 import { toast } from 'react-toastify';
 import EventTypes from './EventTypes';
 import OptionBoxStyle from '../../../components/Payment/OptionBoxStyle';
-import TicketTypeContainer from '../../../components/Payment/TicketTypeContainer';
+import HotelsOptions from './HotelsOptions';
+import TicketSummaryMessage from './TicketSummaryMessage';
 
 export default function Payment() {
   const { enrollment } = useEnrollment();
@@ -59,73 +58,6 @@ export default function Payment() {
           renderEventOrPayment()
       }
     </>
-  );
-}
-
-function HotelsOptions({ ticketTypes, selectedType, setTicketTypeId, ticketTypeId }) {
-  const [hotelType, setHotelType] = useState({});
-  if (Object.keys(selectedType).length === 0) return '';
-
-  return (
-    <HotelOptionsContainer>
-      <h2>Ótimo! Agora escolha sua modalidade de hospedagem</h2>
-      <div>
-        {ticketTypes
-          .filter((type) => type.name === 'Presencial')
-          .map((type) => (
-            <IncludesHotelBox
-              type={type}
-              name={type.includesHotel ? 'Com Hotel' : 'Sem Hotel'}
-              key={type.id}
-              setTicketTypeId={setTicketTypeId}
-              hotelType={hotelType}
-              setHotelType={setHotelType}
-            />
-          ))}
-      </div>
-      {hotelType.name ? (
-        <TicketSummaryMessage selectedOption={hotelType} ticketTypeId={ticketTypeId} />
-      ) : (
-        ''            
-      )}
-    </HotelOptionsContainer>
-  );
-}
-
-function IncludesHotelBox({ type, name, setTicketTypeId, setHotelType, hotelType }) {
-  function handleHotelOption() {
-    setTicketTypeId(type.id);
-    setHotelType({ name, price: type.price });
-  }
-
-  return (
-    <HotelBoxStyle hotelType={hotelType.name ? hotelType.name : ''} name={name} onClick={handleHotelOption}>
-      <h3>{name}</h3>
-      <span>+ R$ {type.price / 100 - 250}</span>
-    </HotelBoxStyle>
-  );
-}
-
-function TicketSummaryMessage({ selectedOption, ticketTypeId }) {
-  const { postCreatedTicket } = useCreateTicket();
-  const formattedPrice = (selectedOption.price / 100).toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' });
-
-  async function reserveTicket() {
-    try {
-      postCreatedTicket({ ticketTypeId });
-      toast('Ticket reservado com sucesso!');
-    } catch (error) {
-      toast('Não foi possível reservar seu ticket!');
-    }
-  }
-
-  return (
-    <Summary>
-      <h2>
-        Fechado! O total ficou em <strong>{formattedPrice}</strong>. Agora é só confirmar:
-      </h2>
-      <Button onClick={reserveTicket}>Reservar Ingresso</Button>
-    </Summary>
   );
 }
 
@@ -226,10 +158,6 @@ const StyledTypography = styled(Typography)`
   margin-bottom: 27px !important;
 `;
 
-const HotelOptionsContainer = styled(TicketTypeContainer)`
-  margin-top: 44px;
-`;
-
 const PaymentContainer = styled.section`
   margin-top: 30px;
   h2 {
@@ -253,20 +181,6 @@ const ConfirmationMessage = styled.div`
   }
 `;
 
-const HotelBoxStyle = styled(OptionBoxStyle)`
-  border: ${(props) => (props.name === props.hotelType ? 'none' : '1px solid #cecece')};
-  background-color: ${(props) => (props.name === props.hotelType ? '#FFEED2' : '#ffffff')};
-`;
-
-const Summary = styled.footer`
-  margin-top: 44px;
-
-  h2 {
-    font-size: 20px;
-    color: #8e8e8e;
-    margin-bottom: 10px;
-  }
-`;
 const PaymentStatusContainer = styled(OptionBoxStyle)`
   width: 290px;
   height: 108px;
